rename case to invokedCase
