+++ api/workAllocation2/util.ts
@@ -4,7 +4,7 @@ import { http } from '../lib/http';
 import { setHeaders } from '../lib/proxy';
 import { TaskPermission, VIEW_PERMISSIONS_ACTIONS_MATRIX } from './constants/actions';
 import { Person, PersonDomain } from './interfaces/person';
-import { Action, Caseworker, CaseworkerApi, Location, LocationApi } from './interfaces/task';
+import { Action, Caseworker, CaseworkerApi, Location, LocationApi } from './interfaces/common';
 
 export function prepareGetTaskUrl(baseUrl: string, taskId: string): string {
   return `${baseUrl}/task/${taskId}`;
