+++ src/work-allocation-2/models/cases/index.ts
@@ -1,13 +1,13 @@
-import InvokedCaseAction from './invoked-case-action.model';
+import Case from './case.model';
 import CaseAction from './case-action.model';
 import CaseFieldConfig from './case-field-config.model';
 import CaseServiceConfig from './case-service-config.model';
-import Case from './case.model';
+import InvokedCaseAction from './invoked-case-action.model';
 
 export {
-  InvokedCaseAction,
   Case,
   CaseAction,
   CaseFieldConfig,
   CaseServiceConfig,
+  InvokedCaseAction
 };
